Update React app for Streetside testing

+++ pages/_app.js
@@ -1,5 +1,6 @@
 import "../src/styles/globals.css";
 import "../src/styles/nprogress.css";
+import "@flaticon/flaticon-uicons/css/all/all.css";
 import { CacheProvider } from "@emotion/react";
 import { Provider as ReduxProvider } from "react-redux";
 import createEmotionCache from "../src/utils/create-emotion-cache";
@@ -22,6 +23,8 @@ import { useTranslation } from "react-i18next";
 import useScrollToTop from "../src/api-manage/hooks/custom-hooks/useScrollToTop";
 import { useEffect } from "react";
 import ModuleChecker from "../src/components/module-select/ModuleChecker";
+import ProSubscriptionExpiredModal from "../src/components/pro-plan/ProSubscriptionExpiredModal";
+import App from "next/app";
 
 Router.events.on("routeChangeStart", nProgress.start);
 Router.events.on("routeChangeError", nProgress.done);
@@ -43,6 +46,7 @@ function MyApp(props) {
     Component,
     emotionCache = clientSideEmotionCache,
     pageProps,
+    initialSettings,
   } = props;
   const getLayout = Component.getLayout ?? ((page) => page);
   const { t } = useTranslation();
@@ -52,12 +56,18 @@ function MyApp(props) {
 
   // Version check
   useEffect(() => {
+    if (!currentVersion) return;
+
     const storedVersion = localStorage.getItem("appVersion");
     if (storedVersion !== currentVersion) {
+      const storedSettings = localStorage.getItem("settings");
       localStorage.clear();
+      if (storedSettings) {
+        localStorage.setItem("settings", storedSettings);
+      }
       localStorage.setItem("appVersion", currentVersion);
     }
-  }, [currentVersion]);
+  }, []);
 
   return (
     <>
@@ -65,7 +75,7 @@ function MyApp(props) {
       <CacheProvider value={emotionCache}>
         <QueryClientProvider client={queryClient}>
           <ReduxProvider store={store}>
-            <SettingsProvider>
+            <SettingsProvider initialSettings={initialSettings}>
               <SettingsConsumer>
                 {(value) => (
                   <ThemeProvider
@@ -77,8 +87,9 @@ function MyApp(props) {
                   >
                     <RTL direction={value?.settings?.direction}>
                       <CssBaseline />
-                      <Toaster position="top-center" />
+                      <Toaster position="top-center" containerStyle={{ zIndex: 99999 }} />
                       <ModuleChecker />
+                      <ProSubscriptionExpiredModal />
                       {getLayout(<Component {...pageProps} />)}
                     </RTL>
                   </ThemeProvider>
@@ -86,7 +97,7 @@ function MyApp(props) {
               </SettingsConsumer>
             </SettingsProvider>
           </ReduxProvider>
-          <ReactQueryDevtools initialIsOpen={false} position="bottom-right" />
+          {/* <ReactQueryDevtools initialIsOpen={false} position="bottom-right" /> */}
         </QueryClientProvider>
       </CacheProvider>
     </>
@@ -95,3 +106,26 @@ function MyApp(props) {
 
 export default MyApp;
 export { getServerSideProps };
+
+const getThemeFromCookieHeader = (cookieHeader = "") => {
+  const match = cookieHeader.match(/(?:^|;\s*)themeMode=(dark|light)(?:;|$)/);
+  return match?.[1] || null;
+};
+
+MyApp.getInitialProps = async (appContext) => {
+  const appProps = await App.getInitialProps(appContext);
+  const cookieSource =
+    appContext?.ctx?.req?.headers?.cookie ||
+    (typeof document !== "undefined" ? document.cookie : "");
+
+  const cookieTheme = getThemeFromCookieHeader(cookieSource);
+
+  return {
+    ...appProps,
+    initialSettings: {
+      direction: "ltr",
+      responsiveFontSizes: true,
+      theme: cookieTheme === "dark" ? "dark" : "light",
+    },
+  };
+};
